#38. Move logic from parent class Numeric to Integer, Float instead. Add russian language for fractional numbers.

+++ lib/numbers_and_words/core_ext/integer.rb
@@ -1,4 +1,4 @@
-class Numeric
+class Integer
   def to_words options = {}
     to_figures_array.to_words NumbersAndWords::Strategies::Base.factory, options
   end
